+++ merlin/common/sample_index.py
@@ -6,7 +6,7 @@
 #
 # LLNL-CODE-797170
 # All rights reserved.
-# This file is part of Merlin, Version: 1.4.1.
+# This file is part of Merlin, Version: 1.5.0.
 #
 # For details, see https://github.com/LLNL/merlin.
 #
@@ -50,9 +50,11 @@ def new_dir(path):
 
 def uniform_directories(num_samples=MAX_SAMPLE, bundle_size=1, level_max_dirs=100):
     """Create a directory hierarchy uniformly stepping up directory sizes."""
-    directory_sizes = [bundle_size * level_max_dirs]
+    directory_sizes = [bundle_size]
     while directory_sizes[0] < num_samples:
         directory_sizes.insert(0, directory_sizes[0] * level_max_dirs)
+    # We've gone over the total number of samples, remove the first entry
+    del directory_sizes[0]
     return directory_sizes