+++ merlin/examples/workflows/hpc_demo/sample_processor.py
@@ -5,6 +5,7 @@ import sys
 
 import pandas as pd
 
+
 def load_samples(sample_file_paths):
     """Read in processed sample files, returning flat list."""
     samples = []
@@ -15,21 +16,21 @@ def load_samples(sample_file_paths):
                 samples.append(sample)
 
     return samples
-                
+
+
 def setup_argparse():
     parser = argparse.ArgumentParser(
         description="Read in and analyze samples from plain text file"
     )
 
     parser.add_argument(
-        "sample_file_paths", help="paths to sample files", default="",
-        nargs='+'
+        "sample_file_paths", help="paths to sample files", default="", nargs="+"
     )
 
-    parser.add_argument("--results",
-                        help="Name of output json file",
-                        default="samples.json")
-    
+    parser.add_argument(
+        "--results", help="Name of output json file", default="samples.json"
+    )
+
     return parser
 
 
@@ -41,7 +42,7 @@ def main():
     samples = load_samples(args.sample_file_paths)
 
     # Count up the occurences
-    namesdf = pd.DataFrame({"Name":samples})
+    namesdf = pd.DataFrame({"Name": samples})
 
     names = namesdf["Name"].value_counts()
 
@@ -53,6 +54,7 @@ def main():
         pathlib.Path(absdir).mkdir(parents=True, exist_ok=True)
 
     names.to_json(args.results)
-    
+
+
 if __name__ == "__main__":
     sys.exit(main())